web_controller/requirements.txt
Flask~=2.1.2
redis~=4.3.3
rq~=1.10.1
minio
python-dotenv